feat: home v2 screen
