+++ client/src/core/constants/feature-toggles.enum.ts
@@ -1,0 +1,3 @@
+export enum FeatureToggles {
+    HomeV2 = 'home-v2'
+}
