+++ client/src/feat/home/home-v1.stories.tsx
@@ -3,7 +3,7 @@ import { Shell } from "core/ui/shell";
 import { withParents } from "core/utils";
 import type { Meta, StoryObj as Story } from "storybook-solidjs";
 import { HomeProviderMock } from "./context";
-import { Home } from "./home";
+import { Home } from "./home-v1";
 import { withI18n } from "./resources";
 
 export default {
